EmotionCardのプレゼンテーションロジックをViewStateに移動
